Fixed block read bug

+++ MP3_Player_Software/Experiment 4/src/Modules/FAT.h
@@ -47,8 +47,8 @@
 uint8_t Read_Sector(uint32_t sector_number, uint16_t sector_size, uint8_t* array_for_data);
 
 uint8_t FAT_read8(uint16_t offset, uint8_t* array_name);
-uint8_t FAT_read16(uint16_t offset, uint8_t* array_name);
-uint8_t FAT_read32(uint16_t offset, uint8_t* array_name);
+uint16_t FAT_read16(uint16_t offset, uint8_t* array_name);
+uint32_t FAT_read32(uint16_t offset, uint8_t* array_name);
 
 uint8_t FAT_mountDrive(uint8_t* array_in);
 uint32_t FAT_getFirstSector(uint32_t cluster_number);
